Added test framework based on Jasmine
The csCompTests project uses the Jasmine library for testing: basically,
all you need to do to run the test is reload the index.html file.
The csCompTests project is compiled to a single file, so you don't need
to add each test specification to the index.html separately.
All you need to do to add your tests is create a new specificationXXX.ts
file, and the tests will be added to the suite.