ENH When feature has no name, create one for it.
In many places, each feature needs a Name property. In case the loaded
feature data does not contain it, create one upon loading so we don't
get errors all over the place because it is not set.

The name is resolved by using either:
1 - The feature's style.nameLabel property
2 - A PropertyType Name and a stringFormat referencing the properties,
e.g. stringFormat = '{street} {housenumber}, {zipcode}'.
3 - The first property that is present.
4 - A GUID otherwise.