ENH iconUri can use stringformatting
When resolving the feature.style.iconUri, check whether we are dealing
with a stringFormatted uri, e.g.
images/LabelledHouse-{energyLabel}.png

In this case, the energyLabel, a property of the feature, is used to
resolve the uri, so this could become when the energyLabel = 4:
images/LabelledHouse-4.png

So for each energyLabel (including the missing label, if any), you need
to supply an image.
Further note that, for the legend, we also need an iconUri that is
independent of the feature: in this case, the brackets are removed, e.g.
images/LabelledHouse-energyLabel.png