ENH TimelineCtrl now loads a few locales (English, Dutch, German, French), and it uses the $layerService.currentLocale to select it..
ENH When configuring the Timeline, I wanted to turn off the event margin (so it became less high). However, I did not want to encode it directly in the directive, so I've added a TimelineService provider to set the timeline options.
See project.ts ITimelineOptions interface for common options. .

REF: Refactored the Timeline's css style to csStyles.css

NOTE When turning off the event timeline, also set the css
#focustimeContainer {
    bottom: 45px;
}