Fix pagination
The angular-utils-pagination, upon fresh checkout, was renamed to angularUtils-pagination, causing errors at runtime due to the fact that the template couldn't be found anymore.
Actually, due to this template being specifed as html, we still need to install the package in the example project, and cannot merge iti in our third party libs.