Adding Wallaby.js testing
Locally, I have the Wallaby.js test plugin working in Code. This allows me to evaluate the tests on-the-fly. I can even see the expected result (in the taskbar). 
As Wallaby seems to only track/use files that are in a sub-folder, I copy the dist-bower output to the test folder.