LegendList updated
Besides the fact that a FeatureType can now contain legend expressions (legendExpr of type PropertyType[]) that show additional features, it is also only updated when the visibility of the legend in the DOM is true.
Also, it compensates for filtered entries, i.e. when a particular feature type is filtered from the map, it means that it won't show up anymore.