FIX filtering bug in initBarFilter
In BarFilterCtrl.initBarFilter, the group filter contained an error computing the bins. 
```Math.floor(a / binWidth) * binWidth;``` 
should have been
```min + Math.floor(a / binWidth) * binWidth;```
By neglecting the minimum, we would always loose the smallest features (expect when the minimum was 0).

In addition, the max needed to be increased too, otherwise we would loose the features with the largest property value.

This fix still needs to be applied to the ScatterFilter too.